The app is divided on 3 modules: model, domain and presentation.Basic frameworks of the madules are done.
