+++ app/build.gradle
@@ -3,6 +3,11 @@ apply plugin: 'com.android.application'
 android {
     compileSdkVersion 22
     buildToolsVersion "21.1.2"
+
+    aaptOptions {
+        useNewCruncher false
+    }
+
     defaultConfig {
         applicationId "ua.od.macra.smartsked"
         minSdkVersion 14
@@ -25,6 +30,7 @@ dependencies {
     compile 'com.android.support:appcompat-v7:22.0.0'
     compile 'com.github.nirhart:parallaxscroll:1.0'
     compile 'javax.json:javax.json-api:1.0'
-    compile 'com.squareup.retrofit:retrofit:1.9.0'
     compile 'com.android.support:support-v4:21.0.2'
+    compile 'com.jakewharton:butterknife:6.1.0'
+    compile 'se.emilsjolander:stickylistheaders:2.6.0'
 }
