+++ app/src/main/AndroidManifest.xml
@@ -1,17 +1,17 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="ua.od.macra.smartsked" >
+    package="ua.od.macra.smartsked">
 
     <uses-permission android:name="android.permission.INTERNET" />
-    <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE"/>
+    <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
 
     <application
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:theme="@style/SmartskedTheme"
-        android:label="@string/app_name" >
+        android:label="@string/app_name">
         <activity
-            android:name="ua.od.macra.smartsked.MainActivity"
+            android:name=".ui.MuinActivity"
             android:label="@string/app_name">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -19,11 +19,11 @@
             </intent-filter>
         </activity>
         <activity
-            android:name="ua.od.macra.smartsked.ShedActivity"
+            android:name=".ui.ShedActivity"
             android:label="@string/app_name">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="ua.od.macra.smartsked.MainActivity" />
+                android:value=".ui.MuinActivity" />
         </activity>
     </application>
