+++ app/src/main/java/ua/od/macra/smartsked/Prefs.java
@@ -1,0 +1,7 @@
+package ua.od.macra.smartsked;
+
+/**
+ * Created by Roodie on 28.04.2015.
+ */
+public class Prefs {
+}
