--- app/src/main/java/ua/od/macra/smartsked/api/entitities/ScheduleData.java
@@ -1,10 +1,0 @@
-package ua.od.macra.smartsked.api.entitities;
-
-import java.util.List;
-import java.util.Map;
-
-/**
- * Created by Roodie on 08.04.2015.
- */
-public class ScheduleData {
-}
