--- app/src/main/java/ua/od/macra/smartsked/api/tasks/FetchScheduleTask.java
@@ -1,7 +1,0 @@
-package ua.od.macra.smartsked.api.tasks;
-
-/**
- * Created by Roodie on 08.04.2015.
- */
-public class FetchScheduleTask {
-}
