+++ common/README.md
@@ -1,0 +1,3 @@
+The module that shares libraries and utilities.
+For this project I have chosen a Message Bus system, which is very useful for Broadcast events, or for establishing  communication between components.
+Using this system the modules have a very low coupling.
