+++ common/build.gradle
@@ -1,0 +1,8 @@
+apply plugin: 'java'
+
+sourceCompatibility = 1.8
+targetCompatibility = 1.8
+
+dependencies {
+    compile 'com.squareup:otto:1.3.5'
+}
