+++ domain/README.md
@@ -1,0 +1,3 @@
+The domain layer is completely independent of the presentation layer, in which will reside business logic of the application.
+This module is totally independent of the Android framework.
+Its dependencies are the model module and common module.
