+++ domain/build.gradle
@@ -1,0 +1,10 @@
+apply plugin: 'java'
+sourceCompatibility = 1.8
+targetCompatibility = 1.8
+
+
+dependencies {
+    compile project(':common')
+    compile project(':model')
+
+}
