+++ domain/src/main/java/com/roodie/domain/Usecase.java
@@ -1,0 +1,9 @@
+package com.roodie.domain;
+
+/**
+ * Created by Roodie on 03.05.2015.
+ */
+public interface Usecase {
+
+    public void execute();
+}
