+++ model/README.md
@@ -1,0 +1,3 @@
+The module will be responsible for providing information, this layer does not know about the domain or  the presentation. This module could implement a connection and a interface with a REST API.
+n this layer there are also implemented the entities of the application.
+Its dependencies only are the common module and for now, the library used to manage requests to the API.
