+++ model/build.gradle
@@ -1,0 +1,8 @@
+apply plugin: 'java'
+sourceCompatibility = 1.8
+targetCompatibility = 1.8
+
+dependencies {
+    compile project(':common')
+    compile 'com.squareup.retrofit:retrofit:1.9.0'
+}
