+++ model/src/main/java/com/roodie/model/entitities/FacultiesWrapper.java
@@ -1,0 +1,19 @@
+package com.roodie.model.entitities;
+
+import java.util.List;
+
+/**
+ * Created by Roodie on 03.05.2015.
+ */
+public class FacultiesWrapper {
+
+    public List<Faculty> faculties;
+
+    public List<Faculty> getFaculties() {
+        return faculties;
+    }
+
+    public void setFaculties(List<Faculty> faculties) {
+        this.faculties = faculties;
+    }
+}
