+++ model/src/main/java/com/roodie/model/entitities/GroupsWrapper.java
@@ -1,0 +1,19 @@
+package com.roodie.model.entitities;
+
+import java.util.List;
+
+/**
+ * Created by Roodie on 03.05.2015.
+ */
+public class GroupsWrapper {
+
+    public List<Group> groups;
+
+    public List<Group> getGroups() {
+        return groups;
+    }
+
+    public void setGroups(List<Group> groups) {
+        this.groups = groups;
+    }
+}
