+++ model/src/main/java/com/roodie/model/entitities/Institute.java
@@ -1,0 +1,46 @@
+package com.roodie.model.entitities;
+
+import com.google.gson.annotations.Expose;
+
+public class Institute {
+    @Expose
+    public Integer id;
+    @Expose
+    public String name;
+    @Expose
+    public String abbreviation;
+
+    public int getId() {
+        return id;
+    }
+
+    public void setId(Integer id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getAbbreviation() {
+        return abbreviation;
+    }
+
+    public void setAbbreviation(String abbreviation) {
+        this.abbreviation = abbreviation;
+    }
+
+    @Override
+    public String toString() {
+        return "Institute{" +
+                "id=" + id +
+                ", name='" + name + '\'' +
+                ", abbreviation='" + abbreviation + '\'' +
+                '}';
+    }
+}
+
