+++ model/src/main/java/com/roodie/model/entitities/InstitutesWrapper.java
@@ -1,0 +1,19 @@
+package com.roodie.model.entitities;
+
+import java.util.List;
+
+/**
+ * Created by Roodie on 03.05.2015.
+ */
+public class InstitutesWrapper {
+
+    public List<Institute> institutes;
+
+    public List<Institute> getInstitutes() {
+        return institutes;
+    }
+
+    public void setInstitutes(List<Institute> institutes) {
+        this.institutes = institutes;
+    }
+}
