+++ model/src/main/java/com/roodie/model/entitities/LessonsInstance.java
@@ -1,0 +1,36 @@
+package com.roodie.model.entitities;
+
+import java.util.List;
+
+/**
+ * Created by Roodie on 08.04.2015.
+ */
+public class LessonsInstance {
+    /// ??????????????????????????
+    private String date;
+    private List<Lesson> lessons;
+
+    public String getDate() {
+        return date;
+    }
+
+    public void setDate(String date) {
+        this.date = date;
+    }
+
+    public List<Lesson> getLessons() {
+        return lessons;
+    }
+
+    public void setLessons(List<Lesson> lessons) {
+        this.lessons = lessons;
+    }
+
+    @Override
+    public String toString() {
+        return "LessonsWrapper{" +
+                "date='" + date + '\'' +
+                ", lessons=" + lessons +
+                '}';
+    }
+}
