+++ model/src/main/java/com/roodie/model/entitities/ScheduleData.java
@@ -1,0 +1,7 @@
+package com.roodie.model.entitities;
+
+/**
+ * Created by Roodie on 08.04.2015.
+ */
+public class ScheduleData {
+}
